refactor(mcp): rename V1BackedByokStore -> DbBackedByokStore, move to stores.py
The BYOK store reads the persisted per-user credentials table (shared with the BYOK entry
flow), not v1 lifecycle code that 1b replaces, so "V1Backed" / v1_adapters.py mis-signaled a
temporary crutch. Move it to stores.py as DbBackedByokStore (the source of record); no
behavior change. v1_adapters.py is reserved for the lifecycle bodies (token_exchange, oauth)
that actually delegate to v1 and get swapped in 1b.
